bdd/requirements.txt
pymongo==4.15.1
python-dotenv==1.0.1